Changed Cooldown data type from double to long to match other time data type.

+++ Ascension/src/blockfighter/server/entities/player/skills/Skill.java
@@ -15,7 +15,7 @@ public abstract class Skill {
 
     public abstract boolean isPassive();
 
-    public abstract double getMaxCooldown();
+    public abstract long getMaxCooldown();
 
     public abstract Byte getReqWeapon();
