+++ Ascension/src/blockfighter/server/entities/player/skills/SkillPassiveVitalHit.java
@@ -9,7 +9,7 @@ public class SkillPassiveVitalHit extends SkillPassive {
     private static final byte SKILL_CODE = Globals.PASSIVE_VITALHIT;
     private static final boolean IS_PASSIVE;
     private static final byte REQ_WEAPON;
-    private static final double MAX_COOLDOWN;
+    private static final long MAX_COOLDOWN;
 
     private static final double BASE_VALUE, MULT_VALUE;
 
@@ -18,7 +18,7 @@ public class SkillPassiveVitalHit extends SkillPassive {
         HashMap<String, Integer> dataHeaders = Globals.getDataHeaders(data, null);
 
         REQ_WEAPON = Globals.loadReqWeapon(data, dataHeaders);
-        MAX_COOLDOWN = Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_MAXCOOLDOWN_HEADER);
+        MAX_COOLDOWN = (long) Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_MAXCOOLDOWN_HEADER);
         BASE_VALUE = Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_BASEVALUE_HEADER);
         MULT_VALUE = Globals.loadDoubleValue(data, dataHeaders, Globals.SKILL_MULTVALUE_HEADER);
         IS_PASSIVE = Globals.loadBooleanValue(data, dataHeaders, Globals.SKILL_PASSIVE_HEADER);
@@ -34,7 +34,7 @@ public class SkillPassiveVitalHit extends SkillPassive {
     }
 
     @Override
-    public double getMaxCooldown() {
+    public long getMaxCooldown() {
         return MAX_COOLDOWN;
     }
